+++ iothub_client/tests/iothubtr_amqp_tel_msgr_ut/iothubtr_amqp_tel_msgr_ut.c
@@ -2059,7 +2059,7 @@ TEST_FUNCTION(telemetry_messenger_create_failure_checks)
         TELEMETRY_MESSENGER_HANDLE handle = telemetry_messenger_create(config, TEST_DEVICE_ID);
 
         // assert
-        sprintf(error_msg, "On failed call %zu", i);
+        sprintf(error_msg, "On failed call %lu", (unsigned long)i);
         ASSERT_IS_NULL(handle, error_msg);
     }
 
@@ -2590,7 +2590,7 @@ TEST_FUNCTION(telemetry_messenger_do_work_create_message_receiver_failure_checks
         telemetry_messenger_do_work(handle);
 
         // assert
-        sprintf(error_msg, "On failed call %zu", i);
+        sprintf(error_msg, "On failed call %lu", (unsigned long)i);
         ASSERT_IS_TRUE(saved_messagereceiver_open_on_message_received == NULL, error_msg);
     }
 
@@ -2645,8 +2645,8 @@ TEST_FUNCTION(telemetry_messenger_do_work_create_message_sender_failure_checks)
         telemetry_messenger_do_work(handle);
 
         // assert
-        sprintf(error_msg, "On failed call %zu", i);
-
+        sprintf(error_msg, "On failed call %lu", (unsigned long)i);
+        
         ASSERT_ARE_EQUAL(int, TELEMETRY_MESSENGER_STATE_STARTING, saved_on_state_changed_callback_previous_state, error_msg);
         ASSERT_ARE_EQUAL(int, TELEMETRY_MESSENGER_STATE_ERROR, saved_on_state_changed_callback_new_state, error_msg);
 
@@ -3154,7 +3154,7 @@ TEST_FUNCTION(telemetry_messenger_send_async_failure_checks)
         int result = telemetry_messenger_send_async(handle, TEST_IOTHUB_MESSAGE_LIST_HANDLE, TEST_on_event_send_complete, TEST_IOTHUB_CLIENT_HANDLE);
 
         // assert
-        sprintf(error_msg, "On failed call %zu", i);
+        sprintf(error_msg, "On failed call %lu", (unsigned long)i);
         ASSERT_ARE_NOT_EQUAL(int, 0, result, error_msg);
     }
 
@@ -3550,7 +3550,7 @@ TEST_FUNCTION(telemetry_messenger_send_message_disposition_failure_checks)
         int result = telemetry_messenger_send_message_disposition(handle, &disposition_info, TELEMETRY_MESSENGER_DISPOSITION_RESULT_ACCEPTED);
 
         // assert
-        sprintf(error_msg, "On failed call %zu", i);
+        sprintf(error_msg, "On failed call %lu", (unsigned long)i);
         ASSERT_ARE_NOT_EQUAL(int, 0, result, error_msg);
     }
 
@@ -3635,7 +3635,7 @@ TEST_FUNCTION(telemetry_messenger_retrieve_options_failure_checks)
         OPTIONHANDLER_HANDLE result = telemetry_messenger_retrieve_options(handle);
 
         // assert
-        sprintf(error_msg, "On failed call %zu", i);
+        sprintf(error_msg, "On failed call %lu", (unsigned long)i);
         ASSERT_IS_NULL(result, error_msg);
     }
